requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.4
coreapi==2.3.3
coreschema==0.0.4
datasets==2.13.0
dill==0.3.6
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
drf-yasg==1.21.6
emoji==2.5.1
evaluate==0.4.0
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.6.0
huggingface-hub==0.15.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
lit==16.0.6
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pyarrow==12.0.1
pyre-extensions==0.0.29
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
responses==0.18.0
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
setfit==0.7.0
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
sympy==1.12
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.2
triton==2.0.0
typing-extensions==4.6.3
typing-inspect==0.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
xformers==0.0.20
xxhash==3.2.0
yarl==1.9.2
